+++ src/main/java/tv/porst/jhexview/IMenuCreator.java
@@ -6,8 +6,7 @@ import javax.swing.JPopupMenu;
 
 /**
  * This interface must be implemented by all classes that want to provide
- * context menus for the JHexView control.
- *
+ * context menus for the {@link JHexView} control.
  */
 public interface IMenuCreator
 {
@@ -15,11 +14,11 @@ public interface IMenuCreator
    * This function is called to generate a popup menu after the user
    * right-clicked somewhere in the hex control.
    *
-   * @param offset
-   *          The current cursor position at the time of the right-click.
+   * @param offset The current caret position at the time of the right-click.
+   *        Will be in range {@code [0; JHexView.getData().getDataLength())}
    *
-   * @return The popup menu suitable for that offset or null if no popup menu
-   *         should be shown.
+   * @return The popup menu suitable for that offset or {@code null} if no popup
+   *         menu should be shown.
    */
   JPopupMenu createMenu(long offset);
 }